Location Service系の処理を分離
