+++ TRViS/DTAC/VerticalTimetableView.LocationService.cs
@@ -1,0 +1,103 @@
+using DependencyPropertyGenerator;
+
+using TRViS.IO.Models;
+using TRViS.Services;
+
+namespace TRViS.DTAC;
+
+[DependencyProperty<bool>("IsRunStarted")]
+[DependencyProperty<bool>("IsLocationServiceEnabled")]
+public partial class VerticalTimetableView : Grid
+{
+	LocationService LocationService { get; } = new();
+
+	partial void OnIsLocationServiceEnabledChanged(bool newValue)
+	{
+		if (newValue)
+			SetNearbyCheckInfo(CurrentRunningRow);
+
+		LocationService.IsEnabled = newValue;
+	}
+
+	private void LocationService_IsNearbyChanged(object? sender, bool oldValue, bool newValue)
+	{
+		if (!IsRunStarted || !IsEnabled || CurrentRunningRow is null)
+			return;
+
+		if (newValue)
+		{
+			SetCurrentRunningRow(NextRunningRow);
+		}
+		else if (CurrentRunningRow is not null)
+		{
+			CurrentRunningRow.LocationState = VerticalTimetableRow.LocationStates.RunningToNextStation;
+
+			SetNearbyCheckInfo(NextRunningRow);
+		}
+	}
+
+	private void SetNearbyCheckInfo(VerticalTimetableRow? nextRunningRow)
+	{
+		if (nextRunningRow?.BindingContext is TimetableRow nextRowData)
+		{
+			LocationService.NearbyCenter
+				= nextRowData.Location is LocationInfo
+				{
+					Latitude_deg: double lat,
+					Longitude_deg: double lon
+				}
+					? new Location(lat, lon)
+					: null;
+
+			LocationService.NearbyRadius_m = nextRowData.Location.OnStationDetectRadius_m ?? 300;
+		}
+	}
+
+	public void SetCurrentRunningRow(int index)
+		=> SetCurrentRunningRow(index, RowViewList.ElementAtOrDefault(index));
+
+	public void SetCurrentRunningRow(VerticalTimetableRow? value)
+		=> SetCurrentRunningRow(value is null ? -1 : RowViewList.IndexOf(value), value);
+
+	void SetCurrentRunningRow(int index, VerticalTimetableRow? value)
+	{
+		if (CurrentRunningRowIndex == index || CurrentRunningRow == value)
+			return;
+
+		if (RowViewList.ElementAtOrDefault(index) != value)
+			throw new ArgumentException("value is not match with element at given index", nameof(value));
+
+		MainThread.BeginInvokeOnMainThread(() =>
+		{
+			if (_CurrentRunningRow is not null)
+				_CurrentRunningRow.LocationState = VerticalTimetableRow.LocationStates.Undefined;
+
+			_CurrentRunningRow = value;
+
+			if (value is not null)
+			{
+				CurrentRunningRowIndex = index;
+				value.LocationState = VerticalTimetableRow.LocationStates.AroundThisStation;
+
+				int rowCount = Grid.GetRow(value);
+
+				Grid.SetRow(CurrentLocationBoxView, rowCount + 1);
+				Grid.SetRow(CurrentLocationLine, rowCount);
+
+				CurrentLocationBoxView.IsVisible = false;
+				CurrentLocationLine.IsVisible = false;
+
+				if (value.LocationState != VerticalTimetableRow.LocationStates.Undefined)
+				{
+					ScrollRequested?.Invoke(this, new(Math.Max(rowCount - 1, 0) * RowHeight.Value));
+				}
+			}
+			else
+				CurrentRunningRowIndex = -1;
+
+			NextRunningRow = RowViewList.ElementAtOrDefault(index + 1);
+
+			SetNearbyCheckInfo(value);
+		});
+	}
+}
